A start on an implementation of XSLT to transform our subset of Docbook to XHTML.
Rationalle for implementing this is:
 * Sanity. Customising Docbook's stock XSLT is a nightmare.
 * Simple, clean markup. This is not a general purpose implementation; we can output exactly the XHTML we need. This makes life easier for CSS, too.
 * Quality of output. With Docbook's stock XSLT's XHTML output, I had to compromise several times due to CSS or XSLT customisation difficulties.
 * Remove a dependency.

Quality was the final straw which caused me to try this approach; I saw that Docbook's stock XSLT output tables with footnotes inside a `<tr>`, causing wide footnotes to stretch the table width. There's no way to override that without reimplementing the whole table-handling code. Why not just output a `<div>` under the table? Good grief.